requirements/dev.txt
-r base.txt
pytest>=9.0.2
pytest-cov>=7.0.0
pytest-mock>=3.15.1
pytest-django>=4.12.0
